.history/requirement_20210214202903.txt
Flask 
Flask-SQLAlchemy
Flask-Script
Flask-restful
Flask-Migrate
python-dotenv
PyMySQL
